requirements.txt
Werkzeug==2.0.1
Flask[async]==2.0.1
flask-cors==4.0.0
pymongo==3.12.0
motor==2.5.1
bcrypt==4.0.1
PyJWT==2.8.0
asgiref==3.7.2
hypercorn==0.14.3
asgiref==3.7.2